Decided how to capture probable chorus of song, and what datapoints to make
Chorus usually starts at around 25% of the song. Search for segments that
start at around this time. Use the closest one. If none exist, the just
start capturing at 25% into the song. Capture until the end of the segment
or 25 seconds into the segment.

With this time segment, we will create a datapoint for every note in the
segment. It has the starting note one-hot encoded, four ints or floats that
describe the change in pitch, the first 3 timbre elements of the first note
and the probable key of song. Each data point will have 20 dimensions.

Looks something like this:
{C?, C#?, D?, D#?, E?, F?, F#?, G?, G#?, A?, A#?, B?,
n0->n1, n1->n2, n2->n3, n3->n4, tim0, tim1, tim2, key}